2- Basic Reflection Systems/basic.py
from typing import List, Sequence
from dotenv import load_dotenv
from langchain_core.messages import BaseMessage, HumanMessage
from langgraph.graph import END, MessageGraph
from chains import generation_chain, reflection_chain

load_dotenv()

REFLECT = "REFLECT"
GENERATE = "GENERATE"
graph = MessageGraph()

def generate_node(state):
    return generation_chain.invoke({
        "messages": state
    })


def reflect_node(messages):
    response = reflection_chain.invoke({
        "messages": messages
    })
    return [HumanMessage(content=response.content)]

graph.add_node(GENERATE, generate_node)
graph.add_node(REFLECT, reflect_node)
graph.set_entry_point(GENERATE)


def should_continue(state):
    print(f"[State length]: {len(state)}")
    if len(state) > 2:
        print("→ Ending the graph.")
        return "END"
    print("→ Reflecting again.")
    return "REFLECT"

graph.add_conditional_edges(
    GENERATE,
    should_continue,
    path_map={
        "REFLECT": REFLECT,
        "END": END
    }
)

graph.add_edge(REFLECT, GENERATE)

app = graph.compile()

print(app.get_graph().draw_mermaid())
app.get_graph().print_ascii()

response = app.invoke(HumanMessage(content="AI Agents taking over content creation"))

print(response)
