requirements.txt
langchain
langchain_community
langchain-google-genai
python-dotenv
langgraph
grandalf
pydantic
langchain_groq
langgraph-checkpoint-sqlite
# groq